remove float32 to float64

+++ somber/ng.py
@@ -92,7 +92,7 @@ class Ng(Base):
         data = json.load(open(path))
 
         weights = data['weights']
-        weights = np.asarray(weights, dtype=np.float32)
+        weights = np.asarray(weights, dtype=np.float64)
 
         s = cls(data['num_neurons'],
                 data['data_dimensionality'],
